another incremental checkin
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@2385 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/rick/opt/api/rexx.h
@@ -295,7 +295,7 @@ typedef RexxReturnCode (REXXENTRY *PFNREXXVARIABLEPOOL)(PSHVBLOCK);
 
 /* This typedef simplifies coding of an External Function.           */
 
-typedef unsigned int REXXENTRY RexxRoutineHandler(const char *,
+typedef size_t REXXENTRY RexxRoutineHandler(const char *,
                                   size_t,
                                   PCONSTRXSTRING,
                                   const char *,
